Change memoryBase handling. Bump to 1.0.9
The data section is set using memoryBase either as an immutable import
or as a constant expression. The previous calculation of the data
start using the memoryBase global is not actually supported by the
spec and implementations are now strict about it. See Initializer
Expressions at https://webassembly.org/docs/modules/